ons: Download Data from the Office for National Statistics

Provides functions to download and tidy statistical data published by the Office for National Statistics https://www.ons.gov.uk. Covers GDP, inflation (CPI, CPIH, RPI), unemployment, employment, wages, trade, retail sales, house prices, productivity, population, and public sector finances. Most series are fetched from the ONS website using its CSV time series endpoint. House price data is sourced from HM Land Registry https://www.gov.uk/government/organisations/land-registry. Data is cached locally between sessions.

See also

Author

Maintainer: Charles Coverdale charlesfcoverdale@gmail.com